zabbix_complex_version/requirements.txt
pyzabbix==1.2.1
atlassian-python-api==3.35.0